[Factoriser outils MCP Verso] refactor(mcp): extract Verso tools into shared module
Extract all 6 Verso-specific MCP tools, 3 prompts, and 1 resource from
both server.ts and stdio.ts into a shared verso-tools.ts module.
Eliminates code duplication and provides a clean extension point for
future UI Kit tools.
