+++ packages/mcp/src/server.ts
@@ -8,23 +8,19 @@ import { cors } from 'hono/cors'
 import { WebSocketServer, type WebSocket } from 'ws'
 import { z } from 'zod'
 
-import { designPagePrompt, designSystemPrompt, refineDesignPrompt } from './verso-prompts'
-
 import {
   ALL_TOOLS,
-  CODEGEN_PROMPT,
   buildComponent,
   createElement,
-  resolveToTree
+  resolveToTree,
 } from '@verso/core'
 
 import type { ParamDef, ParamType } from '@verso/core'
 
+import { ok, fail, registerVersoTools } from './verso-tools'
+
 const require = createRequire(import.meta.url)
 const MCP_VERSION: string = (require('../package.json') as { version: string }).version
-
-type MCPContent = { type: 'text'; text: string } | { type: 'image'; data: string; mimeType: string }
-type MCPResult = { content: MCPContent[]; isError?: boolean }
 
 const RPC_TIMEOUT = 30_000
 
@@ -32,15 +28,6 @@ interface PendingRequest {
   resolve: (value: unknown) => void
   reject: (error: Error) => void
   timer: ReturnType<typeof setTimeout>
-}
-
-function ok(data: unknown): MCPResult {
-  return { content: [{ type: 'text', text: JSON.stringify(data, null, 2) }] }
-}
-
-function fail(e: unknown): MCPResult {
-  const msg = e instanceof Error ? e.message : String(e)
-  return { content: [{ type: 'text', text: JSON.stringify({ error: msg }) }], isError: true }
 }
 
 export function paramToZod(param: ParamDef): z.ZodType {
@@ -272,162 +259,8 @@ export function startServer(options: ServerOptions = {}) {
       )
     }
 
-    // --- Verso-specific MCP tools ---
-
-    register(
-      'get_design_context',
-      {
-        description: 'Get assembled design context from the 4-layer Design Context Engine. Returns principles, trends, task templates, and project context to help the AI produce professional-quality designs.',
-        inputSchema: z.object({
-          scope: z.enum(['full', 'typography', 'colors', 'spacing', 'components', 'accessibility']).optional().describe('Filter context by design domain'),
-          taskDescription: z.string().optional().describe('Description of the design task to classify and get specific guidelines'),
-        })
-      },
-      async (args: Record<string, unknown>) => {
-        try {
-          const { assembleContext } = await import('../../../packages/design-context/src/index.ts')
-          const context = assembleContext({
-            scope: args.scope as string | undefined,
-            taskDescription: args.taskDescription as string | undefined,
-          })
-          return ok(context)
-        } catch (e) { return fail(e) }
-      }
-    )
-
-    register(
-      'validate_design',
-      {
-        description: 'Validate the current design for accessibility, spacing consistency, touch targets, and hardcoded values. Returns issues with severity (critical/warning/info) and suggestions.',
-        inputSchema: z.object({
-          nodeId: z.string().optional().describe('Validate a specific node. If omitted, validates the entire page.'),
-        })
-      },
-      async (args: Record<string, unknown>) => {
-        try {
-          const result = await sendToBrowser({ command: 'tool', args: { name: 'get_page_tree', args: {} } })
-          const res = result as { ok?: boolean; result?: unknown }
-          if (!res.result) return fail(new Error('No design data'))
-          return ok({ message: 'Design validation complete', nodeId: args.nodeId, issues: [] })
-        } catch (e) { return fail(e) }
-      }
-    )
-
-    register(
-      'get_design_guidelines',
-      {
-        description: 'Get universal design principles (Layer 1 of the Design Context Engine): visual hierarchy, Gestalt laws, typography, color theory, spacing, accessibility, psychology, composition.',
-        inputSchema: z.object({})
-      },
-      async () => {
-        try {
-          const { assembleContext } = await import('../../../packages/design-context/src/index.ts')
-          const context = assembleContext({ scope: 'full' })
-          return ok({ principles: context.principles })
-        } catch (e) { return fail(e) }
-      }
-    )
-
-    register(
-      'suggest_structure',
-      {
-        description: 'Get a suggested node tree structure for a given design task. Returns the task template with guidelines, suggested structure, color/typography/spacing guidance, common mistakes, and quality checklist.',
-        inputSchema: z.object({
-          taskDescription: z.string().describe('Description of the design task (e.g. "landing page hero", "dashboard with KPIs", "login form")'),
-        })
-      },
-      async (args: Record<string, unknown>) => {
-        try {
-          const { classifyTask } = await import('../../../packages/design-context/src/classifier.ts')
-          const { getTaskContext } = await import('../../../packages/design-context/src/task-context.ts')
-          const taskType = classifyTask(args.taskDescription as string)
-          const template = getTaskContext(taskType)
-          return ok({ taskType, template })
-        } catch (e) { return fail(e) }
-      }
-    )
-
-    register(
-      'save_as_design',
-      {
-        description: 'Export the current document as a .design file (Verso native JSON format). Returns the JSON content.',
-        inputSchema: z.object({
-          name: z.string().optional().describe('Document name for the .design file'),
-        })
-      },
-      async (args: Record<string, unknown>) => {
-        try {
-          const result = await sendToBrowser({ command: 'tool', args: { name: 'get_page_tree', args: {} } })
-          return ok({ format: 'design', name: args.name ?? 'Untitled', data: result })
-        } catch (e) { return fail(e) }
-      }
-    )
-
-    register(
-      'get_design_prompt',
-      {
-        description: 'Get pixel-precise JSX component templates from the Verso UI Kit. Call BEFORE any render call to get the exact component specs. Returns ready-to-use JSX snippets with correct dimensions, colors, spacing, and typography.',
-        inputSchema: z.object({
-          components: z.array(z.string()).optional().describe('Specific components to get (e.g. ["button-primary", "card", "navbar"]). If omitted, returns ALL component specs.'),
-        })
-      },
-      async (args: Record<string, unknown>) => {
-        try {
-          const { UI_KIT_SPECS, getAllSpecsAsPrompt, getComponentSpec } = await import('./ui-kit-specs.ts')
-          const requested = args.components as string[] | undefined
-
-          if (!requested || requested.length === 0) {
-            return ok({ prompt: getAllSpecsAsPrompt(), availableComponents: Object.keys(UI_KIT_SPECS) })
-          }
-
-          const specs: Record<string, unknown> = {}
-          for (const name of requested) {
-            const spec = getComponentSpec(name)
-            if (spec) specs[name] = spec
-          }
-          return ok({ specs, availableComponents: Object.keys(UI_KIT_SPECS) })
-        } catch (e) { return fail(e) }
-      }
-    )
-
-    // --- End Verso tools ---
-
-    register(
-      'get_codegen_prompt',
-      {
-        description: 'Get design-to-code generation guidelines. Call before generating frontend code.',
-        inputSchema: z.object({})
-      },
-      async () => ok({ prompt: CODEGEN_PROMPT })
-    )
-
-    // --- Verso MCP Prompts ---
-
-    mcpServer.prompt(designPagePrompt.name, designPagePrompt.description, designPagePrompt.params,
-      (args: Record<string, unknown>) => designPagePrompt.handler(args))
-
-    mcpServer.prompt(designSystemPrompt.name, designSystemPrompt.description, designSystemPrompt.params,
-      (args: Record<string, unknown>) => designSystemPrompt.handler(args))
-
-    mcpServer.prompt(refineDesignPrompt.name, refineDesignPrompt.description, refineDesignPrompt.params,
-      (args: Record<string, unknown>) => refineDesignPrompt.handler(args))
-
-    // --- Verso MCP Resource ---
-
-    mcpServer.resource(
-      'design-guidelines',
-      'design://guidelines',
-      { mimeType: 'application/json', description: 'Universal design principles from the Design Context Engine (Layer 1)' },
-      async () => {
-        try {
-          const { assembleContext } = await import('../../../packages/design-context/src/index.ts')
-          const ctx = assembleContext({ scope: 'full' })
-          return { contents: [{ uri: 'design://guidelines', mimeType: 'application/json', text: JSON.stringify(ctx.principles, null, 2) }] }
-        } catch {
-          return { contents: [{ uri: 'design://guidelines', mimeType: 'application/json', text: '{}' }] }
-        }
-      }
-    )
+    // --- Verso tools, prompts & resource (shared module) ---
+    registerVersoTools({ register, sendToBrowser, mcpServer })
 
     const transport = new WebStandardStreamableHTTPServerTransport({
       sessionIdGenerator: () => id
